ref: Make EventMetaData an extractor
