+++ server/src/extractors/mod.rs
@@ -2,8 +2,10 @@ use actix_web::State;
 
 use service::ServiceState;
 
+mod event_meta;
 mod signed_json;
 
+pub use self::event_meta::*;
 pub use self::signed_json::*;
 
 /// An extractor for the entire service state.
